Fixing nits
