+++ .gitignore
@@ -1,1 +1,3 @@
+# Byte-compiled / optimized / DLL files
 /venv
+__pycache__/
