+++ requirements.txt
@@ -44,9 +44,9 @@ singledispatch==3.6.1
 six==1.15.0
 sqlparse==0.4.1
 text-unidecode==1.3
-Twisted==21.2.0
+Twisted
 txaio==21.2.1
 typing-extensions==3.10.0.0
 urllib3==1.26.4
 yarl==1.6.3
-zope.interface==5.4.0
+zope.interface
